Moved REST API tests under freediscovery/server/tests
