+++ freediscovery/server/tests/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+
