+++ freediscovery/server/tests/test_duplicate_detection.py
@@ -1,0 +1,63 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import os
+import pytest
+import json
+import itertools
+from unittest import SkipTest
+from numpy.testing import assert_equal, assert_almost_equal
+
+from .. import fd_app
+from ...utils import _silent, dict2type, sdict_keys
+from ...ingestion import DocumentIndex
+from ...exceptions import OptionalDependencyMissing
+from ...tests.run_suite import check_cache
+
+from .base import parse_res, V01, app, app_notest, get_features, get_features_lsi
+
+#=============================================================================#
+#
+#                     Duplicates detection
+#
+#=============================================================================#
+
+
+@pytest.mark.parametrize('kind, options', [['simhash', {'distance': 3}],
+                                             ['i-match', {}]])
+def test_api_dupdetection(app, kind, options):
+
+    if kind == 'simhash':
+        try:
+            import simhash
+        except ImportError:
+            raise SkipTest
+
+    dsid, pars = get_features(app, hashed=False)
+
+    method = V01 + "/feature-extraction/{}".format(dsid)
+    res = app.get(method)
+    assert res.status_code == 200
+    data = parse_res(res)  # TODO unused variable
+
+    url = V01 + "/duplicate-detection" 
+    pars = { 'parent_id': dsid,
+             'method': kind}
+    res = app.post(url, json=pars)
+    assert res.status_code == 200
+    data = parse_res(res)
+    assert sorted(data.keys()) == sorted(['id'])
+    mid = data['id']
+
+    url += '/{}'.format(mid)
+    res = app.get(url, data=options)
+    assert res.status_code == 200
+    data = parse_res(res)
+    assert sorted(data.keys()) == sorted(['cluster_id'])
+
+    res = app.delete(method)
+    assert res.status_code == 200
